Don't post a GitHub PR URL to a checklist more than once
As per [this feature
request](https://github.com/emmabeynon/github-trello-poster/issues/41) there
are occasions where after a PR URL has been posted to a checklist, a user of
the Trello card may annotate it. In these cases, GitHubTrelloPoster would not
recognise the URL and would post a duplicate. We now use a regex to check if
the PR URL is present on a checklist item, thereby avoiding instances of the PR
being posted more than once.