Check off PR URL items on checklist if PR is closed
As per [this
issue](https://github.com/emmabeynon/github-trello-poster/issues/37) when a PR
is closed but not merged `GitHubTrelloPoster` does not tick this off the
checklist.  After investigation I found that when a PR is either closed or
merged GitHub sets "action" in the payload to "closed". We change app.rb to
check if the "action" is "closed" or not and send this boolean as an argument
to `GitHubPullRequest` upon initialization instead of "merged".