Fix duplicate posts when review requested
When reviews are requested it seems that GitHub sends two payloads - one with
action "opened", one with action "review_requested" - in the same second. This
app doesn't process payloads quickly enough, meaning that there are cases where
two pull request checklists or two of the same PR links are posted. This
overcomes this by checking if the payload has action "review_requested" before
processing it.